Better I/O objects, some shuffling, and small improvements
