+++ gpsdio/cli/env.py
@@ -108,34 +108,3 @@ def schema_cmd(field, indent):
         val = gpsdio.schema.CURRENT
 
     click.echo(json.dumps(_scrub_dict(val), indent=indent, sort_keys=True))
-
-
-
-# @click.command()
-# )
-# @click.pass_context
-# def env(ctx, item, get_help):
-#
-#     """
-#     Information about the gpsdio environment.
-#     """
-#
-#     logger = logging.getLogger('gpsdio-cli-env')
-#     logger.setLevel(ctx.obj['verbosity'])
-#     logger.debug('Starting env')
-#
-#     if item == 'drivers':
-#         for name, driver in gpsdio.drivers._BaseDriver.by_name.items():
-#             click.echo("%s - %s" % (name, driver.io_modes))
-#     elif item == 'compression':
-#         for name, driver in gpsdio.drivers._BaseCompressionDriver.by_name.items():
-#             click.echo("%s - %s" % (name, driver.io_modes))
-#     elif get_help:
-#         if get_help in registered_drivers:
-#             click.echo(registered_drivers[get_help].__doc__)
-#         elif get_help in registered_compression:
-#             click.echo(registered_compression[get_help].__doc__)
-#         else:
-#             click.BadParameter("Unrecognized driver or compression: {}".format(get_help))
-#     else:
-#         raise click.BadParameter('A flag is required.')
